Merging to submit function to edit employee and crud department and position

+++ client/lib/services/auth_service.dart
@@ -1,6 +1,7 @@
 import 'dart:developer';
 
 import 'package:client/services/base_service.dart';
+import 'package:client/utils/api_wrapper.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
@@ -20,10 +21,7 @@ class AuthService extends BaseService {
     final login = await dio.post<Map<String, dynamic>>(
       "/login",
       data: {"email": email, "password": password, "device_name": "android"},
-      options: Options(
-        headers: Map.from({"accept": "application/json"}),
-        validateStatus: (status) => true,
-      ),
+      options: Options(validateStatus: (status) => true),
     );
     if (login.statusCode != 200) {
       log("Error: Login failed: ${login.data?["message"]}");
@@ -62,10 +60,17 @@ class AuthService extends BaseService {
     return null;
   }
 
-  Future<void> logout(BuildContext context) async {
-    final storage = FlutterSecureStorage();
+  Future<ApiResponse> logout() async {
+    final response = await dio.post(
+      "/logout",
+      options: Options(validateStatus: (_) => true),
+    );
+    if (response.statusCode != 200) {
+      log("Error: Logout", error: response.data["message"]);
+      return ApiResponse(message: response.data["message"], success: false);
+    }
+
     await storage.deleteAll();
-    if (!context.mounted) return;
-    context.go("/login");
+    return ApiResponse(message: response.data["message"], success: true);
   }
 }
